ml_assignment/video_detection/requirements.txt
transformers>=4.45.0
torch>=2.2.0
pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0

